v1.4.3.0 Release

See merge request scpcorp/ScPrime!14

+++ Makefile
@@ -1,10 +1,12 @@
-﻿# These variables get inserted into ./build/commit.go
+# These variables get inserted into ./build/commit.go
 BUILD_TIME=$(shell date)
 GIT_REVISION=$(shell git rev-parse --short HEAD)
-GIT_DIRTY=$(shell git diff-index --quiet HEAD -- || echo "✗-")
+GIT_DIRTY=$(shell git diff-index --quiet HEAD -- || echo "modified-")
 
 ldflags= -X gitlab.com/scpcorp/ScPrime/build.GitRevision=${GIT_DIRTY}${GIT_REVISION} \
 -X "gitlab.com/scpcorp/ScPrime/build.BuildTime=${BUILD_TIME}"
+
+GO111MODULE=on
 
 # all will build and install release binaries
 all: release
@@ -19,17 +21,18 @@ cpkg = ./modules/renter
 
 # pkgs changes which packages the makefile calls operate on. run changes which
 # tests are run during testing.
-
 pkgs = ./build \
-	./cmd/sia-node-scanner \
+	./cmd/node-scanner \
 	./cmd/spc \
 	./cmd/spd \
+	./cmd/pubaccess-benchmark \
 	./compatibility \
 	./crypto \
 	./encoding \
 	./modules \
 	./modules/consensus \
 	./modules/explorer \
+	./modules/feemanager \
 	./modules/gateway \
 	./modules/host \
 	./modules/host/contractmanager \
@@ -38,12 +41,14 @@ pkgs = ./build \
 	./modules/renter \
 	./modules/renter/contractor \
 	./modules/renter/filesystem \
+	./modules/stratumminer \
+	./modules/renter/filesystem/siadir \
+	./modules/renter/filesystem/siafile \
 	./modules/renter/hostdb \
 	./modules/renter/hostdb/hosttree \
 	./modules/renter/proto \
-	./modules/renter/siadir \
-	./modules/renter/siafile \
-	./modules/stratumminer \
+	./modules/renter/pubaccessblacklist \
+	./modules/renter/pubaccessportals \
 	./modules/transactionpool \
 	./modules/wallet \
 	./node \
@@ -56,6 +61,7 @@ pkgs = ./build \
 	./siatest/consensus \
 	./siatest/daemon \
 	./siatest/dependencies \
+	./siatest/feemanager \
 	./siatest/gateway \
 	./siatest/host \
 	./siatest/miner \
@@ -65,6 +71,7 @@ pkgs = ./build \
 	./siatest/renterhost \
 	./siatest/transactionpool \
 	./siatest/wallet \
+	./pubaccesskey \
 	./sync \
 	./types \
 	./types/typesutil
@@ -73,12 +80,21 @@ pkgs = ./build \
 # when running a 'make release' command.
 release-pkgs = ./cmd/spc ./cmd/spd
 
+# lockcheckpkgs are the packages that are checked for locking violations.
+lockcheckpkgs = ./modules/renter/hostdb
+
 # run determines which tests run when running any variation of 'make test'.
 run = .
 
 # util-pkgs determine the set of packages that are built when running
 # 'make utils'
-util-pkgs = ./cmd/sia-node-scanner
+util-pkgs = ./cmd/node-scanner ./cmd/pubaccess-benchmark
+
+# dependencies list all packages needed to run make commands used to build, test
+# and lint spc/spd locally and in CI systems.
+dependencies:
+	go get -d ./...
+	./install-dependencies.sh
 
 # fmt calls go fmt on all packages.
 fmt:
@@ -87,31 +103,38 @@ fmt:
 # vet calls go vet on all packages.
 # NOTE: go vet requires packages to be built in order to obtain type info.
 vet:
-	GO111MODULE=on go vet $(pkgs)
-
-lint:
-	GO111MODULE=on go get golang.org/x/lint/golint
-	golint -min_confidence=1.0 -set_exit_status $(pkgs)
-#disable for now
-	#GO111MODULE=on go run ./analysis/cmd/analyze.go -- $(pkgs)
-
-lint-analysis:
-	GO111MODULE=on go run ./analysis/cmd/analyze.go -- $(pkgs)
-
-lint-all:
-	GO111MODULE=on go run ./analysis/cmd/analyze.go -- $(pkgs)
-	golangci-lint run -c .golangci.yml
+	go vet $(pkgs)
 
 # markdown-spellcheck runs codespell on all markdown files that are not
 # vendored.
 markdown-spellcheck:
 	git ls-files "*.md" :\!:"vendor/**" | xargs codespell --check-filenames
 
+# lint runs golangci-lint (which includes golint, a spellcheck of the codebase,
+# and other linters), the custom analyzers, and also a markdown spellchecker.
+lint: markdown-spellcheck lint-analysis
+	golangci-lint run -c .golangci.yml
+
+# lint-ci runs golint.
+lint-ci:
+# golint is skipped on Windows.
+ifneq ("$(OS)","Windows_NT")
+# Linux
+	go get golang.org/x/lint/golint
+	golint -min_confidence=1.0 -set_exit_status $(pkgs)
+endif
+
+# lint-analysis runs the custom analyzers.
+lint-analysis:
+	go run ./analysis/cmd/analyze.go -lockcheck=false -- $(pkgs)
+	go run ./analysis/cmd/analyze.go -lockcheck -- $(lockcheckpkgs)
+
 # spellcheck checks for misspelled words in comments or strings.
-spellcheck:
-	misspell -error .
+spellcheck: markdown-spellcheck
+	golangci-lint run -c .golangci.yml -E misspell
 
 # staticcheck runs the staticcheck tool
+# NOTE: this is not yet enabled in the CI system.
 staticcheck:
 	staticcheck $(pkgs)
 
@@ -129,38 +152,61 @@ dev-race:
 
 # release builds and installs release binaries.
 release:
-	GO111MODULE=on go install -tags='netgo' -ldflags='-s -w $(ldflags)' $(release-pkgs)
+	go install -tags='netgo' -ldflags='-s -w $(ldflags)' $(release-pkgs)
 release-race:
-	GO111MODULE=on go install -race -tags='netgo' -ldflags='-s -w $(ldflags)' $(release-pkgs)
+	go install -race -tags='netgo' -ldflags='-s -w $(ldflags)' $(release-pkgs)
+release-util:
+	go install -tags='netgo' -ldflags='-s -w $(ldflags)' $(release-pkgs) $(util-pkgs)
 
 # clean removes all directories that get automatically created during
 # development.
 clean:
-	rm -rf cover doc/whitepaper.aux doc/whitepaper.log doc/whitepaper.pdf fullcover release 
+ifneq ("$(OS)","Windows_NT")
+# Linux
+	rm -rf cover doc/whitepaper.aux doc/whitepaper.log doc/whitepaper.pdf fullcover release
+else
+# Windows
+# Ignore errors if file doesn't exist
+	- RD /S /Q cover
+	- RD /S /Q doc\whitepaper.aux
+	- RD /S /Q doc\whitepaper.log
+	- RD /S /Q doc\whitepaper.pdf
+	- RD /S /Q fullcover
+	- RD /S /Q release
+endif
 
 test:
-	GO111MODULE=on go test -short -tags='debug testing netgo' -timeout=5s $(pkgs) -run=$(run) -count=$(count)
+	go test -short -tags='debug testing netgo' -timeout=60s $(pkgs) -run=$(run) -count=$(count)
 test-v:
-	GO111MODULE=on go test -race -v -short -tags='debug testing netgo' -timeout=15s $(pkgs) -run=$(run) -count=$(count)
-test-long: clean fmt vet lint
+	go test -race -v -short -tags='debug testing netgo' -timeout=15s $(pkgs) -run=$(run) -count=$(count)
+test-long: clean fmt vet lint-ci
 	@mkdir -p cover
-	GO111MODULE=on go test --coverprofile='./cover/cover.out' -v -race -failfast -tags='testing debug netgo' -timeout=2400s $(pkgs) -run=$(run) -count=$(count)
-test-vlong: clean fmt vet lint
+	go test --coverprofile='./cover/cover.out' -v -race -failfast -tags='testing debug netgo' -timeout=6600s $(pkgs) -run=$(run) -count=$(count)
+
+test-vlong: clean fmt vet lint-ci
+ifneq ("$(OS)","Windows_NT")
+# Linux
 	@mkdir -p cover
-	GO111MODULE=on go test --coverprofile='./cover/cover.out' -v -race -tags='testing debug vlong netgo' -timeout=20000s $(pkgs) -run=$(run) -count=$(count)
+else
+# Windows
+	MD cover
+endif
+	go test --coverprofile='./cover/cover.out' -v -race -tags='testing debug vlong netgo' -timeout=20000s $(pkgs) -run=$(run) -count=$(count)
+
 test-cpu:
-	GO111MODULE=on go test -v -tags='testing debug netgo' -timeout=500s -cpuprofile cpu.prof $(pkgs) -run=$(run) -count=$(count)
+	go test -v -tags='testing debug netgo' -timeout=500s -cpuprofile cpu.prof $(pkgs) -run=$(run) -count=$(count)
 test-mem:
-	GO111MODULE=on go test -v -tags='testing debug netgo' -timeout=500s -memprofile mem.prof $(pkgs) -run=$(run) -count=$(count)
-test-pool:
-	GO111MODULE=on go test -short -parallel=1 -tags='testing debug pool' -timeout=120s ./modules/miningpool -run=$(run)
+	go test -v -tags='testing debug netgo' -timeout=500s -memprofile mem.prof $(pkgs) -run=$(run) -count=$(count)
 bench: clean fmt
-	GO111MODULE=on go test -tags='debug testing netgo' -timeout=500s -run=XXX -bench=$(run) $(pkgs) -count=$(count)
+	go test -tags='debug testing netgo' -timeout=500s -run=XXX -bench=$(run) $(pkgs) -count=$(count)
+test-pool:
+	go test -short -parallel=1 -tags='testing debug pool' -timeout=120s ./modules/miningpool -run=$(run)
+
 cover: clean
 	@mkdir -p cover
 	@for package in $(pkgs); do                                                                                                                                 \
 		mkdir -p `dirname cover/$$package`                                                                                                                      \
-		&& GO111MODULE=on go test -tags='testing debug netgo' -timeout=500s -covermode=atomic -coverprofile=cover/$$package.out ./$$package -run=$(run) || true \
+		&& go test -tags='testing debug netgo' -timeout=500s -covermode=atomic -coverprofile=cover/$$package.out ./$$package -run=$(run) || true \
 		&& go tool cover -html=cover/$$package.out -o=cover/$$package.html ;                                                                                    \
 	done
 
@@ -184,7 +230,7 @@ fullcover: clean
 	@echo "mode: atomic" >> fullcover/fullcover.out
 	@for package in $(pkgs); do                                                                                                                                                             \
 		mkdir -p `dirname fullcover/tests/$$package`                                                                                                                                        \
-		&& GO111MODULE=on go test -tags='testing debug netgo' -timeout=500s -covermode=atomic -coverprofile=fullcover/tests/$$package.out -coverpkg $(cpkg) ./$$package -run=$(run) || true \
+		&& go test -tags='testing debug netgo' -timeout=500s -covermode=atomic -coverprofile=fullcover/tests/$$package.out -coverpkg $(cpkg) ./$$package -run=$(run) || true \
 		&& go tool cover -html=fullcover/tests/$$package.out -o=fullcover/tests/$$package.html                                                                                              \
 		&& tail -n +2 fullcover/tests/$$package.out >> fullcover/fullcover.out ;                                                                                                            \
 	done
@@ -199,4 +245,3 @@ whitepaper:
 	pdflatex -output-directory=doc whitepaper.tex
 
 .PHONY: all fmt install release clean test test-v test-long cover whitepaper
-
